refactor: rename click_ext to options
